requirements.txt
aniso8601==8.0.0
astroid==2.4.2
bcrypt==3.1.7
blinker==1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
oauth2client==4.1.3
oauthlib==3.1.0
pdfkit==0.6.1
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.5.3
pytz==2020.1
reportlab==3.5.44
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
